Redirect to the right place after saving article
For some weird reason, in development mode we were being redirected to a
URL without a port number. Article#url returned with the port number, bur
Article#view_url did not (only in this context, in an automated test it
worked ok).

Since in production mode Noosfero.url_options is guaranteed to return an
empty hash, this has no impact on production servers.