Removing invented %{} 'markup' from links
It turns out that Internet Explorer has problems with it. It looks like it
is trying to parse that text as %-escape caracters, and since the parsing
fails, it does not put those links as proper DOM elements but as 'raw'
objects, and thus the .href attribute does not exist (actually an error
occurs if you try to access such attribute on those objects).

This bug was introduced by the HTTP caching commit (ActionItem1608)

(ActionItem1711)