AI1851: correct the <pre> overflow inside Articles
Solve http://noosfero.org/Development/ActionItem1851

Unhappily, as i see, that is impossible to auto-test if a content is
overflowing a sized box. That is also impossible to auto-test if the
<pre> is bigger then the sized box, despite the content is forcing to
be bigger, and that i can't explain. It's clientWidth shows it with
the parent size. :-(

The good news is that is really easy to correct! :-)
We only need to style white-space: pre-wrap on <pre> elements inside
line-brakes as it must do, and add brakes at the end of the horizontal
space if need (what solves the bug).
http://www.w3.org/TR/CSS21/text.html#propdef-white-space

(ActionItem1851)