Do not send email to comment author
Noosfero sends an email to all admins of a commnunity when a new
comment is posted. This commit removes the author of the comment from
the recipient list.

Additionally, the comment notifier tests were rewritten. The way it was
implemented, the author of the article was always the same as the author
of the comment. Now, these authors are different.