Fix media upload after jQuery upgrade
The solr merge brought a jQuery upgrade from 1.4 to 1.5 (WTF?), and that
broke the jQuery Form plugin. I am adding the very latest form
plugin but jQuery 1.5 is not that recent; that is probably why they
don't seem to interact very well. Note how I had to workaround the
dataType of the response ... weird. :-/