Remove hard dependency on PostgreSQL
In some setups, you want to have the PostgreSQL server in a separate
machine, so we cannot depend on the fact that PostgreSQL is installed in
the same system.

Packages listed in Recommends: will be installed by default, so we don't
break the most common use case that is having both Noosfero and
PostgreSQL in the same system.