Remove useless hotspot
That idea could be useful for streaming comment data, but it is not
useful to notify plugins that a comment was saved before checking e.g.
that it is not SPAM. So this notification must be done by the
CommentHandler class, *after* it decided that the comment received is
valid (e.g. an antispam plugin decided it is not spam).

ActionItem2306