Generic Settings for Plugins
This patch extends the infra developed on the spaminator branch to build
generic settings for plugin by allowing the setting to be instatiated
with any model that supports settings (and respond to the method
settings).

This way a plugin can have any number of settings over any model
instance like a Profile or an Article and not only the Environment.

!! This might conflict with the commit that introduces this idea !!
!! The commit is 7afc68410de26905e442714857effa3e3a13ad1a        !!