Fixed test: features/send_email_to_environment_members.feature
Also removing the :url param on 2 admin view forms that were breaking
the tests. They were originally include to avoid linking to the wrong
page on profiles with own domain.  But this pages will not be visualized
if they are not using the environment domain, therefore there is no need
to used this params.