Allowing plugins to extend cucumber steps and support
Now the plugins can create the folders 'step_definitions' and 'support'
under the folder 'features' to extend cucumber features as it pleases.

The plugins with any extension gain 2 new profiles on the cucumber.yml
file. For example, for mezuro plugin, it will have the 'mezuro' and
'mezuro_selenium' cucumber profiles that will load the appropriate
files. So if you extend cucumber with new steps, hooks, etc; and you
want to run the cucumber tests of mezuro plugin with all these
extensions you must include '-p mezuro' on the cucumber call (or '-p
mezuro-selenium' to run selenium profile). This is already handled by
the rake task that runs the plugins' tests. So it'd be much better to
run 'rake test:noosfero_plugins:mezuro:{cucumber,selenium}' instead.

Also:
   * Moving plugins' 'features' folder to the plugin' root instead of
     inside the folder 'test' (just following Rails convention).
   * Moving specific steps and hooks of mezuro from core to the plugin.
   * Requiring plugin to be enabled to run any of its tests.