[shopping-cart] Fixing double colorbox open and improving requests
We improved some extra requests on the shopping cart, like: visibility
and list actions. They are now being retrieved on the get action all
together.

Also forbidding the user to request more the one action on the cart
simultaneously because the cart in the cookie doesn't supports it.
(We need the session stored in the database!)