find_by_contents: pass the responsability of filter to the engines
Since there are plugins that need to receive the scope clean (that is
without filters applied to it) and since the plugin has the power to not
apply the filter when conveys it (also rewriting the interface to keep
its consistency), I decided to pass the reposability of applying any
filter provided by the core to the engines.

Nevertheless, if the plugin returns nil (that's the same that it would
return if it was disabled) the core will deal with the filters through
the default way (that is without any engine).