container_block: clean test code
Although there are a lot of places on Noosfero where we do this
test_helper ugly inclusion, they are just obsolete. Now we prefer to
just include the test_helper and run the tests through rake tasks or
including the path to the tests helper in the test run (something like:
ruby -Itest ...).