custom-forms-migration: remove/avoid submissions without a form
The migration 20130823151900_associate_fields_to_alternatives.rb from
plugin CustomForms? was crashing due to submissions without a form on
the database. This is a database inconsistence since there shouldn't be
submissions without a form. To fix the problem I'm skipping the on the
migration which may crash with a rescue to allow new tries to execute
properly, creating a new migration to remove the submissions without
form from the database and adding the :dependent => :destroy on form
association with submission to avoid the problem in the future.

Some people bypassed the migration through manual ways. This people
should not worry since the cases which were crashing should not be
running at all.

ActionItem3130