article: author refactoring
I'm adding the author as a record relation instead of calculating it
based on the last_changed_by. This change has a reasonable performance
improvement on pages that access this information, like blog page.

Still saving author_name in order to deal with the problem of the author
being removed.

Including new method to facilitate specific article version retrieval.

ActionItem3201