Fix database upgrade with plugin promoted to base
When a system that has plugin X enabled gets upgraded to a version where
plugin X has been promoted to be a base plugins (i.e. a symlink in
baseplugins/), `rake db:migrate` would blow up with

  Multiple migrations have the version number 12345678901234

Note that we probably don't need to do this deduplication everywhere
else, because the Ruby $LOAD_PATH already works by giving prioriry to
the first path that is found.