publish-content: display only valid options
When the content being displayed is from the user, we should not display
the "publish on my profile" option. The same goes for the "publish on
communities" option if the user has no communities or the "publish on
portal community" option if the environment has no portal community.

If none of this options is available the spread button is also not
displayed.