Added new attribute to Work_assignment article type
Added default email message attribute to be sent when user choose to notify somebody.
Refactored build_email_message method that attaches the URL of the uploaded files to the email message.
Added Java Script to fill in the email field with  an example.
Removed destroy action method on plugin cms controller, now using the existent in the core cms controller.