Support plugin translations in the plugins themselves
When this is merged, the script at
script/move-translations-to-plugins.rb must be executed to move the
translations that are currently in the core to their respective plugins,
i.e. to initialize each plugin .po file with their existing translated
strings, excluding translations that already exist in the core.

Going forward those duplicated translations should be replaced by calls
to c_(), which will do the same as _() but won't be extracted for
translation.