varnish: don't leak HTTP cache headers to clients
This fixes issues when a not-logged-id user logs in, but still sees an
old page. This is because the user's browser has cached the page due to
having received it with 'Cache-Control: public, ...' headers. Now we
make varnish overwrite to always set 'Cache-Control: no-cache' so that
clients will always hit the server again.

For not-logged-in users will also always hit the server, but varnish
will still handle the requests directly for as long as it has a valid
cached version of the page.