permissions: bypass permissions check if the user is admin
If the user is profile/environment admin, bypasses the permissions check
so that he/she always has access to everything related to the profile.

The absence of this feature caused many strange behaviors like env
admins unable to access a community control panel section.

This also caused many complex permission checking codes where we were
forced to test lots of combinations instead of a straight check (e.g.
user is env admin or user is profile admin or user has permission x or
...).

After this change, there might still be redundant testing around the
code, but that might be reviewed slowly.