Fix scrap tests
Tests in scrap_test were failing in some random executions.
It was caused by the way that action_tracker gets the current user when
there is no custom_user setted in track_actions calls.

The current logged in used is obtained with ActionTracker::Record.current_user but it's not setted from unit tests.
The solution is to set the custom_user as the scrap sender, since that it will always be the same user that is logged in when a scrap is created.