dockerfile: add build instructions for development environment
Adding a Dockerfile that builds de development environment. I pushed the
first built machine with this Dockerfile on the following repository:

https://hub.docker.com/r/noosfero/dev/

I'm assuming the responsibility on maintaining this image updated with
the lastest master branch. Whoever feels like help maintaining it send
an email to noosfero-dev requesting access to the repository.