Fix selenium click link web step
Selenium tests were broken in clicking inside the content selection box.

Some HTML structures, such as `<li>` tags nested to `<a>` tags, may not
get along with Selenium and the browser interaction. But are perfectly
fine for the final user.

As we do not want to modify the current behaviour in order to pass
tests, here we handle Selenium's error properly so tests can still pass
without modifying current code.

NOTICE: This only works with selenium-webdriver 2.50.0 and up.