web_steps: retry filling in tinyMCE for up to 5 seconds
Depending on timing trying to fill a tinyMCE control will fail with
`tinymce.editors.article_body is undefined`
(Selenium::WebDriver::Error::JavascriptError), probably because at the
point we try to interact with the editor it is still being loaded by the
browser. Because of that in case of failure we sleep for 1 second and
try again at most 5 times.