Fix random selenium fails when trying to log in
The cause of this issue is still undefined as both steps were leading to
the same page, but with different behaviours.

One possible cause is that clicking the "Login" link should be opening
the log in modal which is not happening. All the taken screenshots are
with the "Login" page displaying. Maybe this is still related to the
previous issues with the selenium click steps addressed by the MRs #780
and #785.