Fix CMSHelper unit test
After creating the ArticleBlock view test at
ff15413b2a8dca45d9f64e57270f294e351a0098 this test got broken.

The root cause was the inclusion of ApplicationHelper for the
article_block_test which led to `custom_options_article` for CMSHelper
tests to fail as the BlogHelper (included by ApplicationHelper) was
overriding the method.

This has been fized by remove the include to ApplicationHelper and
extract the necessary methods to separate helpers which led to two new
methods for ArticleHelper and one new helper the ButtonsHelper.