Extract methods from ApplicationHelper
The method for build profile_image_link was part of ApplicationHelper
but will also be necessary for building proper tests on a upcoming
refactor for PeopleBlockPlugin.

As including ApplicationHelper on unit tests has already caused issues
with unwanted block behaviour regarding their visibility, this is a move
in the direction of avoiding such issues before they happen and allow
reliable future tests.