Remove VideoBlock#content
This method is no longer necessary given the structure at BoxesHelper.
It actually violates MVC as the model is holding view logic.

In order to comply with the BoxesHelper standards, the view got moved
and renamed properly. As well one unit test has been removed as it is
already better covered by the functional ones.