Fixes followers related controllers to use current profile instead of logged in user
Lists and edits followers and circles in control panel of profile that
is being viewed instead of logged in person. Also unfollow person in
control panel based on current viewed profile instead of current logged
in user.

This is necessary because sometimes an environment admin is editing
someone else's profile and those actions should be performed on that
profile instead of admin's profile.