environment: profile kinds feature
This feature allows environment administrators to create custom profile
kinds. These kinds are divided by profile type (Person, Community and
Enterprise). Profiles can define themselves as any of the types created
by the administrator on their creation or editing their informations
later. If the kind is marked as "moderated", a task is created for the
amidnistrator in order to approve a profile to define itself as the
kind in case.

These kinds are now intended to serve as a proper profile
especialization feature instead of using categories, templates or
community memberships for it. They can be used as search
differentiators, style customizations through added css classes and
extra permissions access.