Social Statistics Plugin
This plugin integrates the gem Blazer into Noosfero, providing an
administrative page where the network managers can access usage graphs
of the network. This administrative page is accessible by the user menu
where a new link "Statistics" is added for the environment administrator
when the plugin is enabled. The graphs can be created through the web
interface and are based on sql queries. In this initial version we are
providing the following default graphs:
  * Categories
    * On contents
    * On profiles
  * Contents
    * Through time
    * By category
    * By label
    * By type
  * Labels
    * Through time
    * On contents
    * On profiles
  * Profiles
    * Through time
    * By category
    * By region
    * By label
    * By type
  * Regions
    * On profiles
  * Search Terms
    * On communities
      * By score
      * By occurrences
    * On contents
      * By score
      * By occurrences
    * On people
      * By score
      * By occurrences
    * By score
    * By occurrences

We also implemented support for a limited interface customization
by allowing css, header and footer theme extension.

To customize create a folder named 'blazer' inside your environment
theme folder and use the files:
  * style.css
  * header.html.erb
  * footer.html.erb

The gem currently has no support for translations.